+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- New rule for when the default rules match the example rules provided.
+
 ## [0.8.1] - 2018-08-22
 ### Fixed
 - Rules using action `notEndWith` (like `path-keys-no-trailing-slash`) should not cause validation errors when string length is 1 ([#137])